requirements.txt
ffmpeg-python==0.2.0
python-dotenv==1.0.0
sentence-transformers==4.0.2
supabase==0.7.1
numpy==2.2.4
loguru==0.7.2
markdownify==0.11.6
tqdm==4.67.1
pydantic==1.10.12
pytest==8.3.5
pytest-mock==3.14.0
